requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
Cerberus==1.3.2
click==7.1.2
colorama==0.4.4
Flask==1.1.2
gunicorn==20.0.4
importlib-metadata==3.7.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.1
